Add `components` and `componentsProps` to `PickersCalendarHeader`
- extract `types` into a separate file to avoid API docs generation problems
